more Codecov changes
