+++ contracts/gov/src/tests.rs
@@ -1,10 +1,9 @@
 use crate::contract::{execute, instantiate, migrate, query, reply};
 use crate::error::ContractError;
-use crate::migration::{read_legacy_config, LegacyConfig};
 use crate::mock_querier::mock_dependencies;
 use crate::state::{
-    bank_read, bank_store, config_read, config_store, poll_store, poll_voter_read,
-    poll_voter_store, state_read, Config, Poll, State, TokenManager,
+    bank_read, bank_store, config_read, config_store, legacy_config_read, poll_store,
+    poll_voter_read, poll_voter_store, state_read, Config, LegacyConfig, Poll, State, TokenManager,
 };
 use anchor_token::common::OrderBy;
 use anchor_token::gov::{
@@ -3675,7 +3674,7 @@ fn test_migrate() {
     let info = mock_info(TEST_CREATOR, &coins(2, VOTING_TOKEN));
     let _res = instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
 
-    let legacy_config: LegacyConfig = read_legacy_config(&deps.storage).unwrap();
+    let legacy_config: LegacyConfig = legacy_config_read(&deps.storage).unwrap();
 
     migrate(
         deps.as_mut(),
@@ -3687,39 +3686,19 @@ fn test_migrate() {
     .unwrap();
 
     let new_config: Config = config_read(deps.as_ref().storage).load().unwrap();
-    let expected_config = LegacyConfig {
-        owner: new_config.owner,
-        anchor_token: new_config.anchor_token,
-        quorum: new_config.quorum,
-        threshold: new_config.threshold,
-        voting_period: new_config.voting_period,
-        timelock_period: new_config.timelock_period,
-        expiration_period: new_config.expiration_period,
-        proposal_deposit: new_config.proposal_deposit,
-        snapshot_period: new_config.snapshot_period,
-    };
 
-    assert_eq!(legacy_config.owner, expected_config.owner);
-    assert_eq!(legacy_config.anchor_token, expected_config.anchor_token);
-    assert_eq!(legacy_config.quorum, expected_config.quorum);
-    assert_eq!(legacy_config.threshold, expected_config.threshold);
-    assert_eq!(legacy_config.voting_period, expected_config.voting_period);
-    assert_eq!(
-        legacy_config.timelock_period,
-        expected_config.timelock_period
-    );
+    assert_eq!(legacy_config.owner, new_config.owner);
+    assert_eq!(legacy_config.anchor_token, new_config.anchor_token);
+    assert_eq!(legacy_config.quorum, new_config.quorum);
+    assert_eq!(legacy_config.threshold, new_config.threshold);
+    assert_eq!(legacy_config.voting_period, new_config.voting_period);
+    assert_eq!(legacy_config.timelock_period, new_config.timelock_period);
     assert_eq!(
         legacy_config.expiration_period,
-        expected_config.expiration_period
+        new_config.expiration_period
     );
-    assert_eq!(
-        legacy_config.proposal_deposit,
-        expected_config.proposal_deposit
-    );
-    assert_eq!(
-        legacy_config.snapshot_period,
-        expected_config.snapshot_period
-    );
+    assert_eq!(legacy_config.proposal_deposit, new_config.proposal_deposit);
+    assert_eq!(legacy_config.snapshot_period, new_config.snapshot_period);
     assert_eq!(
         new_config.anchor_voting_escrow,
         deps.api.addr_canonicalize(VOTING_ESCROW).unwrap()
